moves Mentee components out of Headmaster folder into Mentees folder
